+++ instrumentation/opentelemetry-instrumentation-celery/src/opentelemetry/instrumentation/celery/utils.py
@@ -19,6 +19,7 @@ from typing import TYPE_CHECKING, Optional, Tuple
 
 from celery import registry  # pylint: disable=no-name-in-module
 from celery.app.task import Task
+from celery.backends.amqp import Exchange, Queue  # pylint: disable=no-name-in-module
 
 from opentelemetry.semconv._incubating.attributes.messaging_attributes import (
     MESSAGING_MESSAGE_ID,
@@ -123,6 +124,13 @@ def set_attributes_from_context(span, context):
         # set attribute name if not set specially for a key
         if attribute_name is None:
             attribute_name = f"celery.{key}"
+
+        # Serialize Celery classes to the standard python types
+        if isinstance(value, Exchange) or isinstance(value, Queue):
+            value = value.name
+        # Span attributes must be json-serializable
+        elif not isinstance(value, (bool, str, bytes, int, float)):
+            value = str(value)
 
         span.set_attribute(attribute_name, value)